+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-web/src/main/resources/META-INF/resources/js/main.js
@@ -2,19 +2,114 @@ AUI.add(
 	'liferay-workflow-web',
 	function(A) {
 		var WorkflowWeb = {
-			previewBeforeRevert: function(event, renderUrl, namespace, title) {
+
+			openConfirmDeleteDialog : function(title, message, actionUrl) {
+
+				var dialog = Liferay.Util.Window.getWindow(
+					{
+						dialog: {
+							bodyContent: message,
+							destroyOnHide: true,
+							height: 200,
+							resizable: false,
+							toolbars: {
+								footer: [
+									{
+										cssClass: 'btn btn-primary',
+										discardDefaultButtonCssClasses: true,
+										label: Liferay.Language.get('delete'),
+										on: {
+											click: function() {
+												 window.location.assign(actionUrl);
+											}
+										}
+									},
+									{
+										cssClass: 'btn btn-secondary',
+										discardDefaultButtonCssClasses: true,
+										label: Liferay.Language.get('cancel'),
+										on: {
+											click: function() {
+												dialog.destroy();
+											}
+										}
+									}
+								],
+								header: [
+									{
+										cssClass: 'close',
+										discardDefaultButtonCssClasses: true,
+										labelHTML: '<svg class="lexicon-icon" focusable="false"><use data-href="' + Liferay.ThemeDisplay.getPathThemeImages() + '/lexicon/icons.svg#times" /><title>' + Liferay.Language.get('close') + '</title></svg>',
+										on: {
+											click: function(event) {
+												dialog.destroy();
+
+												event.domEvent.stopPropagation();
+											}
+										}
+									}
+								]
+							},
+							width: 600
+						},
+						title: title
+					}
+				);
+
+			},
+
+			previewBeforeRevert: function(event, renderUrl, actionUrl, title) {
 				var instance = this;
 
-				Liferay.Util.Window.getWindow(
+				var form = A.Node.create('<form />');
+
+				form.setAttribute('action', actionUrl);
+				form.setAttribute('method', 'POST');
+
+				var dialog = Liferay.Util.Window.getWindow(
 					{
 						dialog: {
 							destroyOnHide: true,
-							modal: true
+							modal: true,
+							toolbars: {
+								footer: [
+									{
+										cssClass: 'btn btn-secondary',
+										discardDefaultButtonCssClasses: true,
+										label: Liferay.Language.get('cancel'),
+										on: {
+											click: function() {
+												dialog.destroy();
+											}
+										}
+									},
+									{
+										cssClass: 'btn btn-primary',
+										discardDefaultButtonCssClasses: true,
+										label: Liferay.Language.get('restore'),
+										on: {
+											click: function() {
+												submitForm(form);
+											}
+										}
+									}
+								],
+								header: [
+									{
+										cssClass: 'close',
+										discardDefaultButtonCssClasses: true,
+										labelHTML: '<svg class="lexicon-icon" focusable="false"><use data-href="' + Liferay.ThemeDisplay.getPathThemeImages() + '/lexicon/icons.svg#times" /><title>' + Liferay.Language.get('close') + '</title></svg>',
+										on: {
+											click: function(event) {
+												dialog.destroy();
+
+												event.domEvent.stopPropagation();
+											}
+										}
+									}
+								]
+							}
 						},
-						dialogIframe: {
-							bodyCssClass: 'dialog-with-footer'
-						},
-						id: namespace + 'previewBeforeRevert',
 						title: title,
 						uri: renderUrl
 					}