+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseBuild.java
@@ -661,7 +661,7 @@ public abstract class BaseBuild implements Build {
 
 	@Override
 	public Long getLatestStartTimestamp() {
-		Long latestStartTimestamp = getStartTimestamp();
+		Long latestStartTimestamp = getStartTime();
 
 		if (latestStartTimestamp == null) {
 			return null;
@@ -763,7 +763,11 @@ public abstract class BaseBuild implements Build {
 	}
 
 	@Override
-	public Long getStartTimestamp() {
+	public Long getStartTime() {
+		if (startTime != null) {
+			return startTime;
+		}
+
 		JSONObject buildJSONObject = getBuildJSONObject("timestamp");
 
 		if (buildJSONObject == null) {
@@ -772,11 +776,11 @@ public abstract class BaseBuild implements Build {
 
 		long timestamp = buildJSONObject.getLong("timestamp");
 
-		if (timestamp == 0) {
-			return null;
+		if (timestamp != 0) {
+			startTime = timestamp;
 		}
 
-		return timestamp;
+		return startTime;
 	}
 
 	@Override
@@ -1028,8 +1032,8 @@ public abstract class BaseBuild implements Build {
 						message, "jenkins", "Build Reinvoked",
 						reinvokeRule.notificationRecipients);
 				}
-				catch (
-					InterruptedException | IOException | TimeoutException e) {
+				catch (InterruptedException | IOException |
+					   TimeoutException e) {
 
 					throw new RuntimeException(
 						"Unable to send reinvoke notification", e);
@@ -1694,11 +1698,28 @@ public abstract class BaseBuild implements Build {
 			Dom4JUtil.getNewElement(
 				cellElementTagName, null,
 				Dom4JUtil.getNewAnchorElement(
-					getBuildURL() + "testReport", "Test Report")),
-			Dom4JUtil.getNewElement(
-				cellElementTagName, null,
-				JenkinsResultsParserUtil.toDateString(
-					new Date(getStartTimestamp()))),
+					getBuildURL() + "testReport", "Test Report")));
+
+		getStartTime();
+
+		if (startTime == null) {
+			Dom4JUtil.addToElement(
+				buildInfoElement,
+				Dom4JUtil.getNewElement(
+					cellElementTagName, null, "",
+					getJenkinsReportTimeZoneName()));
+		}
+		else {
+			Dom4JUtil.addToElement(
+				buildInfoElement,
+				Dom4JUtil.getNewElement(
+					cellElementTagName, null,
+					toJenkinsReportDateString(
+						new Date(startTime), getJenkinsReportTimeZoneName())));
+		}
+
+		Dom4JUtil.addToElement(
+			buildInfoElement,
 			Dom4JUtil.getNewElement(
 				cellElementTagName, null,
 				JenkinsResultsParserUtil.toDurationString(getDuration())));
@@ -1755,6 +1776,10 @@ public abstract class BaseBuild implements Build {
 		}
 
 		return tableRowElements;
+	}
+
+	protected String getJenkinsReportTimeZoneName() {
+		return _jenkinsReportTimeZoneName;
 	}
 
 	protected Set<String> getJobParameterNames() {
@@ -2216,6 +2241,21 @@ public abstract class BaseBuild implements Build {
 		}
 	}
 
+	protected String toJenkinsReportDateString(Date date, String timeZoneName) {
+		Properties buildProperties = null;
+
+		try {
+			buildProperties = JenkinsResultsParserUtil.getBuildProperties();
+		}
+		catch (IOException ioe) {
+			throw new RuntimeException("Unable to get build properties", ioe);
+		}
+
+		return JenkinsResultsParserUtil.toDateString(
+			date, buildProperties.getProperty("jenkins.report.date.format"),
+			timeZoneName);
+	}
+
 	protected void writeArchiveFile(String content, String path)
 		throws IOException {
 
@@ -2267,6 +2307,7 @@ public abstract class BaseBuild implements Build {
 	protected String repositoryName;
 	protected List<SlaveOfflineRule> slaveOfflineRules =
 		SlaveOfflineRule.getSlaveOfflineRules();
+	protected Long startTime;
 	protected long statusModifiedTime;
 	protected Element upstreamJobFailureMessageElement;
 
@@ -2283,7 +2324,7 @@ public abstract class BaseBuild implements Build {
 			}
 
 			_duration = topLevelBuild.getDuration();
-			_startTimestamp = topLevelBuild.getStartTimestamp();
+			_startTime = topLevelBuild.getStartTime();
 			_timeline = new TimelineDataPoint[size];
 
 			for (int i = 0; i < size; i++) {
@@ -2295,9 +2336,14 @@ public abstract class BaseBuild implements Build {
 		}
 
 		protected void addTimelineData(BaseBuild build) {
-			int endIndex = _getIndex(
-				build.getStartTimestamp() + build.getDuration());
-			int startIndex = _getIndex(build.getStartTimestamp());
+			Long buildStartTime = build.getStartTime();
+
+			if (buildStartTime == null) {
+				return;
+			}
+
+			int endIndex = _getIndex(buildStartTime + build.getDuration());
+			int startIndex = _getIndex(buildStartTime);
 
 			_timeline[startIndex]._invocationsCount++;
 
@@ -2338,8 +2384,7 @@ public abstract class BaseBuild implements Build {
 
 		private int _getIndex(long timestamp) {
 			int index =
-				(int)((timestamp - _startTimestamp) *
-					_timeline.length / _duration);
+				(int)((timestamp - _startTime) * _timeline.length / _duration);
 
 			if (index >= _timeline.length) {
 				return _timeline.length - 1;
@@ -2353,7 +2398,7 @@ public abstract class BaseBuild implements Build {
 		}
 
 		private final long _duration;
-		private final long _startTimestamp;
+		private final long _startTime;
 		private final TimelineDataPoint[] _timeline;
 
 		private static class TimelineDataPoint {
@@ -2393,6 +2438,22 @@ public abstract class BaseBuild implements Build {
 
 	private static final String[] _HIGH_PRIORITY_CONTENT_FLAGS =
 		{"compileJSP", "SourceFormatter.format", "Unable to compile JSPs"};
+
+	private static final String _jenkinsReportTimeZoneName;
+
+	static {
+		Properties properties = null;
+
+		try {
+			properties = JenkinsResultsParserUtil.getBuildProperties();
+		}
+		catch (IOException ioe) {
+			throw new RuntimeException("Unable to get build properties", ioe);
+		}
+
+		_jenkinsReportTimeZoneName = properties.getProperty(
+			"jenkins.report.time.zone");
+	};
 
 	private int _buildNumber = -1;
 	private String _consoleText;