+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TopLevelBuild.java
@@ -788,8 +788,8 @@ public class TopLevelBuild extends BaseBuild {
 			"div", null,
 			Dom4JUtil.getNewElement(
 				"p", null, "Start Time: ",
-				JenkinsResultsParserUtil.toDateString(
-					new Date(getStartTimestamp()))),
+				toJenkinsReportDateString(
+					new Date(getStartTime()), getJenkinsReportTimeZoneName())),
 			Dom4JUtil.getNewElement(
 				"p", null, "Build Time: ",
 				JenkinsResultsParserUtil.toDurationString(getDuration())),