+++ portal-impl/src/com/liferay/portal/model/impl/LayoutTypePortletImpl.java
@@ -84,7 +84,6 @@ import java.util.Set;
  * @author Berentey Zsolt
  * @author Jorge Ferrer
  * @author Raymond Augé
- * @author Neil Griffin
  */
 public class LayoutTypePortletImpl
 	extends LayoutTypeImpl implements LayoutTypePortlet {
@@ -106,13 +105,6 @@ public class LayoutTypePortletImpl
 	public void addModeConfigPortletId(String portletId) {
 		removeModesPortletId(portletId);
 		setModeConfig(StringUtil.add(getModeConfig(), portletId));
-	}
-
-	public void addModeCustomPortletId(String portletId, String portletMode) {
-		removeModesPortletId(portletId);
-		setModeCustom(
-			StringUtil.add(getModeCustom(portletMode), portletId), portletMode);
-		_addedCustomPortletMode = portletMode;
 	}
 
 	@Override
@@ -240,8 +232,7 @@ public class LayoutTypePortletImpl
 			list.addAll(startPortlets);
 		}
 
-		for (int i = 0; i < portlets.size(); i++) {
-			Portlet portlet = portlets.get(i);
+		for (Portlet portlet : portlets) {
 
 			// Add the portlet if and only if it is not also a static portlet
 
@@ -257,10 +248,6 @@ public class LayoutTypePortletImpl
 		}
 
 		return list;
-	}
-
-	public String getAddedCustomPortletMode() {
-		return _addedCustomPortletMode;
 	}
 
 	@Override
@@ -336,9 +323,7 @@ public class LayoutTypePortletImpl
 
 		List<String> columns = getColumns();
 
-		for (int i = 0; i < columns.size(); i++) {
-			String columnId = columns.get(i);
-
+		for (String columnId : columns) {
 			portlets.addAll(getAllPortlets(columnId));
 		}
 
@@ -418,10 +403,6 @@ public class LayoutTypePortletImpl
 		return getTypeSettingsProperty(LayoutTypePortletConstants.MODE_CONFIG);
 	}
 
-	public String getModeCustom(String portletMode) {
-		return getTypeSettingsProperty("mode-" + portletMode);
-	}
-
 	@Override
 	public String getModeEdit() {
 		return getTypeSettingsProperty(LayoutTypePortletConstants.MODE_EDIT);
@@ -486,9 +467,7 @@ public class LayoutTypePortletImpl
 
 		List<Portlet> portlets = new ArrayList<>(portletIds.size());
 
-		for (int i = 0; i < portletIds.size(); i++) {
-			String portletId = portletIds.get(i);
-
+		for (String portletId : portletIds) {
 			Portlet portlet = PortletLocalServiceUtil.getPortletById(
 				getCompanyId(), portletId);
 
@@ -544,12 +523,6 @@ public class LayoutTypePortletImpl
 	@Override
 	public boolean hasModeConfigPortletId(String portletId) {
 		return StringUtil.contains(getModeConfig(), portletId);
-	}
-
-	public boolean hasModeCustomPortletId(
-		String portletId, String portletMode) {
-
-		return StringUtil.contains(getModeCustom(portletMode), portletId);
 	}
 
 	@Override
@@ -1033,9 +1006,7 @@ public class LayoutTypePortletImpl
 
 		List<String> columns = getColumns();
 
-		for (int i = 0; i < columns.size(); i++) {
-			String columnId = columns.get(i);
-
+		for (String columnId : columns) {
 			if (isCustomizable() && isColumnDisabled(columnId)) {
 				continue;
 			}
@@ -1224,10 +1195,6 @@ public class LayoutTypePortletImpl
 	public void setModeConfig(String modeConfig) {
 		setTypeSettingsProperty(
 			LayoutTypePortletConstants.MODE_CONFIG, modeConfig);
-	}
-
-	public void setModeCustom(String modeCustom, String portletMode) {
-		setTypeSettingsProperty("mode-" + portletMode, modeCustom);
 	}
 
 	@Override
@@ -1819,9 +1786,7 @@ public class LayoutTypePortletImpl
 
 		List<String> columns = layoutTemplate.getColumns();
 
-		for (int i = 0; i < columns.size(); i++) {
-			String columnId = columns.get(i);
-
+		for (String columnId : columns) {
 			if (hasNonstaticPortletId(columnId, portletId)) {
 				return true;
 			}
@@ -2039,7 +2004,6 @@ public class LayoutTypePortletImpl
 
 	private static final Layout _nullLayout = new LayoutImpl();
 
-	private String _addedCustomPortletMode;
 	private boolean _customizedView;
 	private final Format _dateFormat =
 		FastDateFormatFactoryUtil.getSimpleDateFormat(